Implement TextShelHost for macOS.

+++ BuildMacOSPackages.sh
@@ -23,6 +23,17 @@ if [[ ! -d "./Packages/$VERSION" ]]; then
     fi
 fi
 
+# Build native libraries
+echo " " 
+echo "Build native libraries"
+echo " "
+cd ./TextShellHost
+. BuildMacOSNativeLibraries.sh
+if [ "$?" != "0" ]; then
+    exit
+fi
+cd ..
+
 # Build packages
 for i in "${!RID_LIST[@]}"; do
     RID=${RID_LIST[$i]}
@@ -31,6 +42,18 @@ for i in "${!RID_LIST[@]}"; do
     echo " " 
     echo "[$PUB_PLATFORM ($RID)]"
     echo " "
+
+    # Build TextShellHost
+    rm -r ./TextShellHost/bin/$CONFIG/net7.0/$RID
+    dotnet clean TextShellHost
+    dotnet restore TextShellHost
+    if [ "$?" != "0" ]; then
+        exit
+    fi
+    dotnet publish TextShellHost -p:SelfContained=true -p:PublishTrimmed=true -p:RuntimeIdentifier=$RID
+    if [ "$?" != "0" ]; then
+        exit
+    fi
 
     # clean
     rm -r ./$APP_NAME/bin/$CONFIG/net7.0/$RID
